Added a review section with tinder like swipe cards, Fixed layout issues
I added a review section which enables a user to swipe pictures of the
drinks and sandwiches that they like/don`t like. It works the same way
tinder does. If the user likes the product they swipe right and if not
they swipe left. I also fixed layout issues on my videos page. The
Youtube videos are no correctly scaled for mobile phones.