Connection to database
I have got my application connected to the firebase API which is were my
database is stored. My app is now synced with the database and is
pulling all the necessary data about my products for it. I can now start
implementing the more complicated functions of my application. I have
already created a better order now page which will calculate how much
the order will be and then offer the user an option to preview the order
before sending it.