Order Now and euro symbols
The order now page is now working. If the user clicks on what they want
it will be added to a preview order page. The price will be displayed at
the bottom of the page along with the amount of items the customer has
ordered. They can then delete what they have ordered from the preview
page. once the items have been submitted a message will be displayed to
the user saying that the order has been sent. The order is not being
stored in my database which I need to fix. Euro symbols have now
replaced the default dollar signs on the page.